cvs: 1.12.13
The reason to go through the effort of including this in the dupe
directory? Xcode 5 Developer Preview no longer has "cvs" included.

This is pretty much a verbatim copy of what you'd get from Xcode 4,
using the same patches and configure arguments from
http://www.opensource.apple.com/

I discovered ENV.j1 due to weird, strange installation failures if
I didn't have it. So, since this is an ANCIENT piece of code, I'm
just going to let that be, and never try to figure out why it doesn't
parallelize properly.

Closes #201.

Signed-off-by: Xiyue Deng <manphiz@gmail.com>