Final bit of polishing to caveats for Yosemite
In Yosemite (and possibly earlier, but I am unsure), the changes needed
for keychain support do not take effect until launchd has reloaded the
changed service .plist.  This can be done via a restart or issuing the
commands I added to the caveats.  It can be done without administrative
privileges.

Closes #372.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>